+++ push_swap/utils/function_Libft.c
@@ -6,7 +6,7 @@
 /*   By: mstasiak <mstasiak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/24 16:27:52 by mstasiak          #+#    #+#             */
-/*   Updated: 2025/02/17 11:17:45 by mstasiak         ###   ########.fr       */
+/*   Updated: 2025/02/19 12:28:52 by mstasiak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,11 +59,11 @@ void	ft_putnbr_fd(int n, int fd)
 		write(fd, &str[length--], 1);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
